src/requirements.txt
azure-storage-blob==12.14.1
boto3~=1.26.16
botocore~=1.29.16
eventlet==0.30.2
Flask==2.2.2
Flask-Mail==0.9.1
Flask-Security-Too==5.0.2
Flask-SocketIo==5.3.2
Flask-SqlAlchemy==3.0.2
libsass==0.22.0
minio==7.1.12
paramiko~=2.12.0
Pillow~=9.3.0
prefect~=2.6.9
psycopg2==2.9.5
pysftp==0.2.9
python-dotenv~=0.21.0
pyzbar==0.1.9
simple-websocket==0.9.0
SQLAlchemy~=1.4.44
tqdm~=4.64.1
watchdog==2.1.9
Werkzeug==2.2.2
WTForms==3.0.1
